feat: move input utility functions to a separate InputUtils.kt file
